created a randomized SampleAsteroid class to test the FlyingObject's collision method

+++ src/main/java/com/example/asteroids2/FlyingObject.java
@@ -17,7 +17,7 @@ public class FlyingObject {
 
     public FlyingObject(int positionX, int positionY,
                         int[][] cornerCoordinates,
-                        int angleWithXAxis, int speed, Team team) {
+                        int angleWithXAxis, double speed, Team team) {
         this(
                 createShape(positionX, positionY, cornerCoordinates, angleWithXAxis),
                 speed, team
